Reorganiza definición de repositorio en package y reorganiza config.yml
